Simplify Experiment Manager
- Update the ExperimentManager methods to consume experiment config
as refactored by the MPIC application wherein a `features` property
contains variant names as keys with a values containing a nested
array holding the values which are mapped to the feature bucket names.

- Update ExperimentManager::enrollUser() to output 3 keys: `enrolled`
which tracks which experiments the user is enrolled in, `assigned`
which tracks which buckets the user has been assigned to, and
`features` which maps the correlation between experiment names and
the feature variants being tested in a given experiment.

The format of the experiments array has been refactored to look like:
```
[
  "enrolled" => [
    "experiment-name-1",
    "experiment-name-2"
  ],
  "assigned" => [
    "sticky_header" => true,
    "homepage_module" => false
  ],
  "features" => [
    "experiment-name-1" => [
      "sticky_header",
      "table_of_contests
    ],
    "experiment-name-2" => [
      "homepage_module"
    ]
  ]
]
```
- Note that the features key allows for 1-to-many relationship
between experiments and feature variants.
- Update relevant tests.

Note: this patch depends on a merge request in the MPIC repo:
https://gitlab.wikimedia.org/repos/data-engineering/mpic/-/merge_requests/135

Bug: T375906
Change-Id: I72a5ced435c12af67888143eac10fc87ee40a8fe