InstrumentsConfigsFetcher: Improve cache hit rate
After creating the MetricsPlatform Grafana dashboard [0], it became
clear that the MPIC request rate is too high at ~3 requests/min rather
than ~1 request/min.

This patch begins to addresses this issue by using the global keyspace
for the cache key. The global keyspace is shared by all wikis within the
same wiki farm. This is a natural fit for storing instrument/experiment
configs, which are for all wikis

[0]: https://grafana.wikimedia.org/d/d1ea56fb-efc3-4f90-9cfe-618d8611c1a4/metricsplatform?orgId=1

Change-Id: Ieb208adea95f06047e5f3e5596b0b0cf8e1e8463