InstrumentConfigsFetcher: Fix use of StatusFormatter for debug logging
StatusFormatter::getWikiText() formats the message in current user's
language, which is not appropriate for PSR-3 debug logging, and which
can't be used in entry points that don't support user sessions
(T388068).

Use StatusFormatter::getPsr3MessageAndContext() instead, which is
intended for this use case, formats the message in English and
with normalized parameters, and can be used anywhere.

Bug: T388068
Change-Id: I680a0287815b39c93ec7fcd84ba66cb91bb7b0d2