xLab: Refactor enrollment logic/enroll on initialization
This change (1) refactors experiment enrollment logic and (2) changes
the time at which experiment enrollment sampling occurs during the
MediaWiki app lifecycle.

Re. (1):

We introduce the Enrollment\EnrollmentAuthorityInterface, which
represents a process that performs experiment enrollment sampling, the
act of enrolling a user into one or more active experiments.

We then extract the logic in ExperimentManager into three distinct
implementations of that interface:

1. EveryoneExperimentsEnrollmentAuthority, which allows xLab to handle
   everyone experiment enrollments from Varnish

2. LoggedInEnrollmentAuthority, which represents the process by which
   logged-in users are enrolled into logged-in experiments

3. OverridesEnrollmentAuthority, which facilitates the experiment
   enrollment override mechanism provided to developers and QTEs by the JS
   xLab SDK to

We then introduce the composite [0] EnrollmentAuthority implementation,
which simply defers to the three implementations above in order.

Finally, we introduce:

* EnrollmentRequest, which adapts the application state for the
  `EnrollmentAuthorityInterface` implementations to use

* EnrollmentResponseBuilder, which accumulates information about active
  experiments and experiment enrollments emitted by the
  `EnrollmentAuthorityInterface` implementations

Re (2).

We move the experiment enrollment sampling procedure to as early as
possible in the MediaWiki application lifecycle so that
`ExperimentManager` is ready to be used by experiment developers as
early as possible.

It's worth noting that (1) and (2) combined allow us to radically
simplify the `ExperimentManager` API for experiment developers also.

Bug: T391732
Change-Id: I87a94621feed79a2b5587ca525ed478c64b71faf