enrollment: Always send subject ID/sampling unit for logged-in experiments
When a user is enrolled in an everyone experiment, we always send a
subject ID of "awaiting" and a sampling unit of "edge-unique" to the
browser.

Past:

We would only send the subject ID and a sampling unit of "mw-user" if
the user was enrolled in the logged-in experiment.  However, if the user
set an override for the logged-in experiment, the subject ID and
sampling unit still wouldn't be set.

Present:

If the user sets an override, then the system will act as if there were
an active experiment and the user will be enrolled into it, and we will
always send a subject ID of "overridden" and a sampling unit of
"mw-user."

Future:

We might consider changing the sampling unit to "overridden" as,
strictly speaking, no experiment enrollment sampling was performed.

Bug: T395777
Change-Id: I19c44c7f51aae8973244475a1aed267885010f57