ExperimentManager: Fix #getExperiment() when uninitialized
What?

Update ExperimentManager#getExperiment() to handle the case where the
experiment manager hasn't been initialized.

Why?

When ExperimentManager is constructed it is initialized with an empty
list of experiment enrollment results. However,
ExperimentManager#getExperiment() doesn't handle this case as it expects
the list of experiment enrollment results to be populated. This case
will occur when the BeforeInitialize hook hasn't run, e.g. API requests
and ResourceLoader requests.

It is out of scope of this change to alter when
ExperimentManager#initialize() is called, which affects what is in scope
for experimentation.

Bug: T401294
Change-Id: I1936bdb33bfb68fd7ffffbf825aab091ba2a472b