ext.xLab: Log events to console for overridden experiments
Why?

Currently, if a user has overridden an experiment, then the SDK will
override certain parameters on all analytics events related to that
experiment, e.g.  we set .experiment.coordinator to "forced" . We can
detect these events and exclude them during analysis.

A point that we tripped on during the initial implementation of this is
that those events would arrive at the server for logged-in experiments
but not for logged-out experiments. What we didn't see at the time is
that this isn't actually an issue because we filter them out during
analysis anyway!

Experiment Platform and several client teams agreed that, if a user has
overridden an experiment, then any analytics events related to that
experiment should be logged to the console.

What?

* Introduce the UnenrolledExperiment and OverriddenExperiment classes,
  which model the user not being enrolled in an experiment and
  overriding their enrollment for an experiment, respectively. This
  makes testing behavior for each case easier and neatens up the code

* Hide UnenrolledExperiment and OverriddenExperiment from the
  public-facing documentation as they both extend Experiment

* Make Experiment an ES6 class

Bug: T402369
Change-Id: Ia36355ba1c1d00129e02d82074ff61ae00924f50