LoggedInExperimentsEnrollmentAuthority: make central ID retrieval more reliable
`centralIdFromName` does not require to local account to already be
attached to the central account, which is often not yet the case, so it
should reliably return the correct central id.

Akin to Icbe639a10cf1a86644a892c0c96047ed29366625, this should make
the enrolment of recently registered logged in users possible in
hooks like LocalUserCreated.

Bug: T401308
Change-Id: I94686a57b0fa637e0d17db267099d542a5e0749d