Add A/B test data to structured logging
An additional field now gets added to structured logging.

For logging that occurs prior to the invocation of onBeforeInitialize
the field may not be present, but for the most part logging of interest
will occur for any onBeforeInitialize callees after MetricsPlatform's
hook implementation runs or other lifecycle callees after that.

Bug: T404262

Change-Id: I70bade89b82df36999b50d0ba8ab02a1f06a6cc3