XLAB_STREAMS -> $wgMetricsPlatformExperimentStreamNames
Why?

The number of streams far outweighs the number of streams used for
experiments. For this reason, xLab maintains an allowlist of streams
whose configs will be sent to the browser as part of the ext.xLab
ResourceLoader module.

To date, however, two teams have needed to use a new stream for their
experiment and in both cases, developers have been confused as to why
their experiments aren't sending events (and maybe even why setting
$wgEventLoggingStreamNames to falsy makes it work).

What?

* Introduce the $wgMetricsPlatformExperimentStreamNames and initialize
  it to [ "product_metrics.web_base" ]

* Update XLab\ResourceLoader\Hooks::getStreamConfigs() to use
  $wgMetricsPlatformExperimentStreamNames

* Update XLab\ResourceLoader\Hooks::getStreamConfigs() and
  ::getStreamConfigsForInstruments() to use the
  EventStreamConfigs.StreamConfigs service so that
  $wgEventLoggingStreamNames doesn't affect their output

* Extract the extant stream config reduction logic to
  XLab\ResourceLoader\Hooks::getMinimumUsableStreamConfig() so that it
  can be reused

* Make the JS version of Experiment#setStream() log a warning when the
  stream isn't registered

Bug: T406332
Depends-On: I99d090dca5f7eb0ec52518d315564b93c9b1f39d
Change-Id: I7fe65c267f0a5526e032d05b5ba284f095c99145