[PHP][JS] `getExperiment()`: Modify the message when the experiment is not active
`getExperiment()` function (PHP and JS) logged a message as a
warning for developers when the current user wasn't enrolled in the
corresponding experiment, when that experiment wasn't active or doesn't exist. The message was a bit confusing because it mentioned that the experiment might not be registered yet.

This patch changes that messages to be more precise

Bug: T412057
Change-Id: I7ee1f2cf7beed1248b5288811270af51a25b101c